added game screen and updated snapshots
