+++ src/components/artist-question-screen/artist-question-screen.jsx
@@ -1,7 +1,6 @@
 import React, {PureComponent} from 'react';
 import {func} from 'prop-types';
 import {artistQuestionPropTypes} from '../../types';
-import Header from '../header/header';
 import {PAUSE_CLASS} from '../../const';
 
 class ArtistQuestionScreen extends PureComponent {
@@ -35,45 +34,41 @@ class ArtistQuestionScreen extends PureComponent {
     const {song, answers} = question;
 
     return (
-      <section className="game game--artist">
-        <Header />
-
-        <section className="game__screen">
-          <h2 className="game__title">Кто исполняет эту песню?</h2>
-          <div className="game__track">
-            <div className="track">
-              <button
-                className="track__button track__button--play"
-                type="button"
-                onClick={this._handlePlayButtonClick}
-              ></button>
-              <div className="track__status">
-                <audio src={song.src}></audio>
-              </div>
+      <section className="game__screen">
+        <h2 className="game__title">Кто исполняет эту песню?</h2>
+        <div className="game__track">
+          <div className="track">
+            <button
+              className="track__button track__button--play"
+              type="button"
+              onClick={this._handlePlayButtonClick}
+            ></button>
+            <div className="track__status">
+              <audio src={song.src}></audio>
             </div>
           </div>
+        </div>
 
-          <form className="game__artist">
-            {answers.map((answer, i) => (
-              <div className="artist" key={answer.artist}>
-                <input
-                  className="artist__input visually-hidden"
-                  type="radio" name="answer"
-                  value={`artist-${i}`}
-                  id={`answer-${i}`}
-                  onChange={(evt) => {
-                    evt.preventDefault();
-                    onAnswer(question, answer);
-                  }}
-                />
-                <label className="artist__name" htmlFor={`answer-${i}`}>
-                  <img className="artist__picture" src={answer.picture} alt={answer.artist} />
-                  {answer.artist}
-                </label>
-              </div>
-            ))}
-          </form>
-        </section>
+        <form className="game__artist">
+          {answers.map((answer, i) => (
+            <div className="artist" key={answer.artist}>
+              <input
+                className="artist__input visually-hidden"
+                type="radio" name="answer"
+                value={`artist-${i}`}
+                id={`answer-${i}`}
+                onChange={(evt) => {
+                  evt.preventDefault();
+                  onAnswer(question, answer);
+                }}
+              />
+              <label className="artist__name" htmlFor={`answer-${i}`}>
+                <img className="artist__picture" src={answer.picture} alt={answer.artist} />
+                {answer.artist}
+              </label>
+            </div>
+          ))}
+        </form>
       </section>
     );
   }
