--- src/components/header/header.test.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import renderer from 'react-test-renderer';
-import Header from './header';
-
-describe(`Header`, () => {
-  it(`should render Header correctly`, () => {
-    const tree = renderer.create(
-        <Header />
-    ).toJSON();
-    expect(tree).toMatchSnapshot();
-  });
-});
